migrate to using .opendistro_security index, but not listen for changes properly
This commit adds a whitelisted_settings.yml file, and modifies the securityadmin.sh script to add the yml file contents to the index.
Also adds a rest endpoint for reading the current whitelisting settings and modifying them. However, when the setting is modified,
the rest filter is unable to detect changes. Need to fix this.